tornando o token invalido caso o usuario seja deletado

+++ src/Service/DeleteUserService.ts
@@ -1,13 +1,15 @@
 import { getRepository } from 'typeorm';
 
 import User from '../models/User';
+import Token from '../models/Token';
 
 interface Request {
     email: string;
+    token?: string;
 }
 
 class DeleteUserService {
-    public async execute({ email }: Request): Promise<void> {
+    public async execute({ email, token }: Request): Promise<void> {
         
         if(!email) {
             throw new Error('please, send email for delete.');
@@ -25,7 +27,25 @@ class DeleteUserService {
             throw new Error('user not found');
         }
 
+        const tokenRepository = getRepository(Token);
+        
+        const tokenInBlackList = await tokenRepository.findOne({
+            hash: token,
+            isvalid: false
+        });
+
+        if(tokenInBlackList) {
+            throw new Error('invalid token')
+        }
+
+        const newToken = tokenRepository.create({
+            hash: token,
+            isvalid: false
+        });
+
+        await tokenRepository.save(newToken);        
         await usersRepository.remove(userExists);
+        
     }
 }
